+++ src/App.tsx
@@ -3,7 +3,6 @@ import './App.css'
 import ToastContainer from './components/common/toast/ToastContainer'
 import { useAuthInit } from './hooks/useAuthInit'
 import { routesConfig } from './routes/routes'
-import NotificationToastContainer from './components/common/toast/NotificationToastContainer'
 
 function App() {
   const { isAuthCheck, isLoading } = useAuthInit() // 인증 초기화
@@ -15,7 +14,6 @@ function App() {
   return (
     <>
       <ToastContainer />
-      <NotificationToastContainer />
       {routes}
     </>
   )
