+++ src/utils/showNotificationToast.tsx
@@ -1,0 +1,33 @@
+import { toast } from 'sonner'
+import NotificationToast from '../components/common/toast/NotificationToast'
+
+export const showChatToast = (
+  title: string,
+  user: string,
+  content: string,
+  date: string,
+  type: 'chat' | 'notice'
+) => {
+  console.log('토스트 호출')
+  toast.custom(
+    (t) => {
+      console.log('토스트 호출2')
+      return (
+        <div className="fixed top-100 z-[999]">
+          <NotificationToast
+            id={t}
+            title={title}
+            user={user}
+            content={content}
+            date={date}
+            type={type}
+          />
+        </div>
+      )
+    },
+    {
+      duration: 5000,
+      unstyled: true,
+    }
+  )
+}
